Use createSlice from RTK for filters
