+++ src/helpers/get-mock-store.ts
@@ -2,7 +2,7 @@ import { configureMockStore } from '@jedmao/redux-mock-store';
 import { RootState } from '../store/store';
 import { initialState as guitarsInitialState } from '../store/guitars/reducer';
 import { initialState as paginationInitialState } from '../store/pagination/reducer';
-import { initialState as filtersInitialState } from '../store/filters/reducer';
+import { initialState as filtersInitialState } from '../store/filters/slice';
 
 export const mockStore = configureMockStore<RootState>();
